It's working, but
there is some problem with mentions by channel times, the are to be integers and now the are strings for whatever reason. On 40 line datatype is Integer(Long), but when map is created on 46 line, the data is converted into string. That is a strange behaivior, since map can handle diffrent data types at once.